Rename Backup-file, new course-backup, new test-data file
